+++ feature/login/src/main/java/dev/enesky/feature/login/signin/SignInScreen.kt
@@ -19,10 +19,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.enesky.core.common.utils.Logger
 import dev.enesky.core.common.utils.ObserveAsEvents
-import dev.enesky.core.design_system.common.ButtonWithImageAndText
-import dev.enesky.core.design_system.login.EmailAuthComponent
-import dev.enesky.core.design_system.login.LineWithTextMiddle
-import dev.enesky.core.design_system.login.LoginHeader
+import dev.enesky.core.design_system.components.ButtonWithImageAndText
+import dev.enesky.core.ui.components.login.EmailAuthComponent
+import dev.enesky.core.ui.components.login.LineWithTextMiddle
+import dev.enesky.core.ui.components.login.LoginHeader
 import dev.enesky.core.design_system.theme.DoodleTheme
 import dev.enesky.core.ui.annotation.PreviewUiMode
 import dev.enesky.feature.login.R
