+++ feature/login/src/main/java/dev/enesky/feature/login/signup/SignUpScreen.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.enesky.core.common.utils.Logger
 import dev.enesky.core.common.utils.ObserveAsEvents
-import dev.enesky.core.design_system.login.EmailAuthComponent
-import dev.enesky.core.design_system.login.LoginHeader
+import dev.enesky.core.ui.components.login.EmailAuthComponent
+import dev.enesky.core.ui.components.login.LoginHeader
 import dev.enesky.core.design_system.theme.DoodleTheme
 import dev.enesky.core.ui.annotation.PreviewUiMode
 import dev.enesky.feature.login.R
